Move WdsTroubleshooter
