+++ src/pages/workspaces/workspace/Data.js
@@ -14,7 +14,6 @@ import { DataTableSaveVersionModal, DataTableVersion, DataTableVersions } from '
 import { getRootTypeForSetTable, ReferenceDataDeleter, ReferenceDataImporter } from 'src/components/data/data-utils';
 import FileBrowser from 'src/components/data/FileBrowser';
 import LocalVariablesContent from 'src/components/data/LocalVariablesContent';
-import { WdsTroubleshooter } from 'src/components/data/WdsTroubleshooter';
 import { icon, spinner } from 'src/components/icons';
 import { ConfirmedSearchInput, DelayedSearchInput } from 'src/components/input';
 import { MenuButton } from 'src/components/MenuButton';
@@ -28,6 +27,7 @@ import { renderDataCell } from 'src/data/data-table/entity-service/renderDataCel
 import { useSavedColumnSettings } from 'src/data/data-table/entity-service/SavedColumnSettings';
 import { EntityUploader } from 'src/data/data-table/shared/EntityUploader';
 import WDSContent from 'src/data/data-table/wds/WDSContent';
+import { WdsTroubleshooter } from 'src/data/data-table/wds/WdsTroubleshooter';
 import { Ajax } from 'src/libs/ajax';
 import { EntityServiceDataTableProvider } from 'src/libs/ajax/data-table-providers/EntityServiceDataTableProvider';
 import { resolveWdsApp, WdsDataTableProvider, wdsProviderName } from 'src/libs/ajax/data-table-providers/WdsDataTableProvider';
